+++ packages/link/__tests__/link.web.test.js
@@ -1,4 +1,4 @@
 import test from "./link-helper";
 import Link, { TextLink } from "../link.web.js";
 
-test(Link, TextLink);
+test(Link, TextLink, "p");