+++ packages/author-head/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.6.7"></a>
+## [0.6.7](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.6...@times-components/author-head@0.6.7) (2017-09-26)
+
+
+
+
+<a name="0.6.6"></a>
+## [0.6.6](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.5...@times-components/author-head@0.6.6) (2017-09-25)
+
+
+### Bug Fixes
+
+* package.json main ([ed150b8](https://github.com/newsuk/times-components/commit/ed150b8))
+
+
+
+
+<a name="0.6.5"></a>
+## [0.6.5](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.3...@times-components/author-head@0.6.5) (2017-09-25)
+
+
+
+
+<a name="0.6.4"></a>
+## [0.6.4](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.3...@times-components/author-head@0.6.4) (2017-09-25)
+
+
+
+
 <a name="0.6.3"></a>
 ## [0.6.3](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.2...@times-components/author-head@0.6.3) (2017-09-25)