+++ packages/provider/CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.8.2"></a>
+## [0.8.2](https://github.com/newsuk/times-components/compare/@times-components/provider@0.8.1...@times-components/provider@0.8.2) (2017-09-26)
+
+
+
+
+<a name="0.8.1"></a>
+## [0.8.1](https://github.com/newsuk/times-components/compare/@times-components/provider@0.8.0...@times-components/provider@0.8.1) (2017-09-25)
+
+
+### Bug Fixes
+
+* package.json main ([ed150b8](https://github.com/newsuk/times-components/commit/ed150b8))
+
+
+
+
+<a name="0.8.0"></a>
+# [0.8.0](https://github.com/newsuk/times-components/compare/@times-components/provider@0.6.3...@times-components/provider@0.8.0) (2017-09-25)
+
+
+### Features
+
+* author profile article links ([#244](https://github.com/newsuk/times-components/issues/244)) ([2d48af7](https://github.com/newsuk/times-components/commit/2d48af7))
+
+
+
+
+<a name="0.7.0"></a>
+# [0.7.0](https://github.com/newsuk/times-components/compare/@times-components/provider@0.6.3...@times-components/provider@0.7.0) (2017-09-25)
+
+
+### Features
+
+* author profile article links ([#244](https://github.com/newsuk/times-components/issues/244)) ([2d48af7](https://github.com/newsuk/times-components/commit/2d48af7))
+
+
+
+
 <a name="0.6.3"></a>
 ## [0.6.3](https://github.com/newsuk/times-components/compare/@times-components/provider@0.6.2...@times-components/provider@0.6.3) (2017-09-25)